+++ package/example/lib/advanced_example/main.dart
@@ -4,6 +4,7 @@ import 'package:easy_form_example/advanced_example/screens/custom_field/custom_f
 import 'package:easy_form_example/advanced_example/screens/login_demo/login_demo_screen.dart';
 import 'package:easy_form_example/advanced_example/screens/save_indicator/save_indicator_screen.dart';
 import 'package:easy_form_example/advanced_example/screens/save_error/save_error_screen.dart';
+import 'package:easy_form_example/advanced_example/screens/clear_value/clear_value_screen.dart';
 
 void main() {
   runApp(MyApp());
@@ -62,6 +63,11 @@ class StartScreen extends StatelessWidget {
                   _DemoButton(
                     text: 'Fields errors',
                     builder: (context) => FieldsErrorsScreen(),
+                  ),
+                  const SizedBox(height: 16),
+                  _DemoButton(
+                    text: 'Clear value',
+                    builder: (context) => ClearValueScreen(),
                   ),
                 ],
               ),